Moved Drawer logic from MainListActivity to separate class
Helper class MainListDrawerHelper handle everything connected with
Drawer.
Small other improvements.
